Move default spartan.el into its own file
Previously, if a spartan.el file did not exists, then init.el wrote a large
string that contained spartan.el. However, since the code was in a string,
this made it cumbersome to edit any changes in the default spartan.el. Moved
the contents of the string to its own file spartan-defaults.el and then,
if spartan.el does not exists, init.el will copy spartan-defaults.el to
spartan.el.